Do not use ID but Id.

+++ oss-dao/src/main/java/de/openschoolserver/dao/TestUserPK.java
@@ -50,8 +50,8 @@ public class TestUserPK implements Serializable {
 	public int hashCode() {
 		final int prime = 31;
 		int hash = 17;
-		hash = hash * prime + this.testId;
-		hash = hash * prime + this.userId;
+		hash = hash * prime + (int) this.testId;
+		hash = hash * prime + (int) this.userId;
 		
 		return hash;
 	}
